support presto to read stream data
