+++ integration/presto/src/main/java/org/apache/carbondata/presto/CarbondataPageSourceProvider.java
@@ -17,27 +17,12 @@
 
 package org.apache.carbondata.presto;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Set;
 
 import static java.util.Objects.requireNonNull;
 
-import org.apache.carbondata.common.CarbonIterator;
-import org.apache.carbondata.core.datastore.block.TableBlockInfo;
-import org.apache.carbondata.core.metadata.AbsoluteTableIdentifier;
 import org.apache.carbondata.core.metadata.schema.table.CarbonTable;
-import org.apache.carbondata.core.scan.executor.QueryExecutor;
-import org.apache.carbondata.core.scan.executor.QueryExecutorFactory;
-import org.apache.carbondata.core.scan.expression.Expression;
-import org.apache.carbondata.core.scan.model.QueryModel;
-import org.apache.carbondata.core.scan.result.iterator.AbstractDetailQueryResultIterator;
-import org.apache.carbondata.core.util.CarbonTimeStatisticsFactory;
-import org.apache.carbondata.hadoop.CarbonInputSplit;
-import org.apache.carbondata.hadoop.CarbonMultiBlockSplit;
-import org.apache.carbondata.hadoop.CarbonProjection;
-import org.apache.carbondata.hadoop.api.CarbonTableInputFormat;
-import org.apache.carbondata.presto.impl.CarbonLocalMultiBlockSplit;
 import org.apache.carbondata.presto.impl.CarbonTableCacheModel;
 import org.apache.carbondata.presto.impl.CarbonTableReader;
 
@@ -45,7 +30,6 @@ import static org.apache.carbondata.presto.Types.checkType;
 
 import com.facebook.presto.hive.HdfsEnvironment;
 import com.facebook.presto.hive.HiveClientConfig;
-import com.facebook.presto.hive.HiveColumnHandle;
 import com.facebook.presto.hive.HivePageSourceFactory;
 import com.facebook.presto.hive.HivePageSourceProvider;
 import com.facebook.presto.hive.HiveRecordCursorProvider;
@@ -57,14 +41,9 @@ import com.facebook.presto.spi.ConnectorSplit;
 import com.facebook.presto.spi.SchemaTableName;
 import com.facebook.presto.spi.connector.ConnectorTransactionHandle;
 import com.facebook.presto.spi.type.TypeManager;
-import com.google.common.collect.ImmutableList;
 import com.google.inject.Inject;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.mapred.JobConf;
-import org.apache.hadoop.mapred.TaskAttemptContextImpl;
-import org.apache.hadoop.mapred.TaskAttemptID;
-import org.apache.hadoop.mapreduce.TaskType;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -103,122 +82,11 @@ public class CarbondataPageSourceProvider extends HivePageSourceProvider {
         new HdfsEnvironment.HdfsContext(session, carbonSplit.getDatabase(), carbonSplit.getTable()),
         new Path(carbonSplit.getSchema().getProperty("tablePath")));
     configuration = carbonTableReader.updateS3Properties(configuration);
-    CarbonDictionaryDecodeReadSupport readSupport = new CarbonDictionaryDecodeReadSupport();
-    PrestoCarbonVectorizedRecordReader carbonRecordReader =
-        createReader(carbonSplit, columns, readSupport, configuration);
-    return new CarbondataPageSource(carbonRecordReader, columns);
-  }
-
-  /**
-   * Create vector reader using the split.
-   */
-  private PrestoCarbonVectorizedRecordReader createReader(HiveSplit carbonSplit,
-      List<? extends ColumnHandle> columns, CarbonDictionaryDecodeReadSupport readSupport,
-      Configuration conf) {
-    QueryModel queryModel = createQueryModel(carbonSplit, columns, conf);
-    if (carbonTableReader.config.getPushRowFilter() == null ||
-        carbonTableReader.config.getPushRowFilter().equalsIgnoreCase("false")) {
-      queryModel.setDirectVectorFill(true);
-      queryModel.setPreFetchData(false);
-    }
-    QueryExecutor queryExecutor =
-        QueryExecutorFactory.getQueryExecutor(queryModel, new Configuration());
-    try {
-      CarbonIterator iterator = queryExecutor.execute(queryModel);
-      readSupport.initialize(queryModel.getProjectionColumns(), queryModel.getTable());
-      PrestoCarbonVectorizedRecordReader reader =
-          new PrestoCarbonVectorizedRecordReader(queryExecutor, queryModel,
-              (AbstractDetailQueryResultIterator) iterator, readSupport);
-      reader.setTaskId(Long.parseLong(carbonSplit.getSchema().getProperty("index")));
-      return reader;
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to create reader ", e);
-    }
-  }
-
-  /**
-   * @param carbondataSplit
-   * @param columns
-   * @return
-   */
-  private QueryModel createQueryModel(HiveSplit carbondataSplit,
-      List<? extends ColumnHandle> columns, Configuration conf) {
-
-    try {
-      CarbonProjection carbonProjection = getCarbonProjection(columns);
-      CarbonTable carbonTable = getCarbonTable(carbondataSplit, conf);
-      conf.set(CarbonTableInputFormat.INPUT_SEGMENT_NUMBERS, "");
-      String carbonTablePath = carbonTable.getAbsoluteTableIdentifier().getTablePath();
-      CarbonTableInputFormat
-          .setTransactionalTable(conf, carbonTable.getTableInfo().isTransactionalTable());
-      CarbonTableInputFormat.setTableInfo(conf, carbonTable.getTableInfo());
-      conf.set(CarbonTableInputFormat.INPUT_DIR, carbonTablePath);
-      conf.set("query.id", queryId);
-      JobConf jobConf = new JobConf(conf);
-      CarbonTableInputFormat carbonTableInputFormat = createInputFormat(jobConf, carbonTable,
-          PrestoFilterUtil.parseFilterExpression(carbondataSplit.getEffectivePredicate()),
-          carbonProjection);
-      TaskAttemptContextImpl hadoopAttemptContext =
-          new TaskAttemptContextImpl(jobConf, new TaskAttemptID("", 1, TaskType.MAP, 0, 0));
-      CarbonMultiBlockSplit carbonInputSplit = CarbonLocalMultiBlockSplit
-          .convertSplit(carbondataSplit.getSchema().getProperty("carbonSplit"));
-      QueryModel queryModel =
-          carbonTableInputFormat.createQueryModel(carbonInputSplit, hadoopAttemptContext);
-      queryModel.setQueryId(queryId);
-      queryModel.setVectorReader(true);
-      queryModel.setStatisticsRecorder(
-          CarbonTimeStatisticsFactory.createExecutorRecorder(queryModel.getQueryId()));
-
-      List<TableBlockInfo> tableBlockInfoList =
-          CarbonInputSplit.createBlocks(carbonInputSplit.getAllSplits());
-      queryModel.setTableBlockInfos(tableBlockInfoList);
-      return queryModel;
-    } catch (IOException e) {
-      throw new RuntimeException("Unable to get the Query Model ", e);
-    }
-  }
-
-  /**
-   * @param conf
-   * @param carbonTable
-   * @param filterExpression
-   * @param projection
-   * @return
-   */
-  private CarbonTableInputFormat<Object> createInputFormat(Configuration conf,
-      CarbonTable carbonTable, Expression filterExpression, CarbonProjection projection) {
-
-    AbsoluteTableIdentifier identifier = carbonTable.getAbsoluteTableIdentifier();
-    CarbonTableInputFormat format = new CarbonTableInputFormat<Object>();
-    try {
-      CarbonTableInputFormat
-          .setTablePath(conf, identifier.appendWithLocalPrefix(identifier.getTablePath()));
-      CarbonTableInputFormat
-          .setDatabaseName(conf, identifier.getCarbonTableIdentifier().getDatabaseName());
-      CarbonTableInputFormat
-          .setTableName(conf, identifier.getCarbonTableIdentifier().getTableName());
-    } catch (Exception e) {
-      throw new RuntimeException("Unable to create the CarbonTableInputFormat", e);
-    }
-    CarbonTableInputFormat.setFilterPredicates(conf, filterExpression);
-    CarbonTableInputFormat.setColumnProjection(conf, projection);
-
-    return format;
-  }
-
-  /**
-   * @param columns
-   * @return
-   */
-  private CarbonProjection getCarbonProjection(List<? extends ColumnHandle> columns) {
-    CarbonProjection carbonProjection = new CarbonProjection();
-    // Convert all columns handles
-    ImmutableList.Builder<HiveColumnHandle> handles = ImmutableList.builder();
-    for (ColumnHandle handle : columns) {
-      handles.add(checkType(handle, HiveColumnHandle.class, "handle"));
-      carbonProjection.addColumn(((HiveColumnHandle) handle).getName());
-    }
-    return carbonProjection;
+    CarbonTable carbonTable = getCarbonTable(carbonSplit, configuration);
+    boolean isDirectVectorFill = carbonTableReader.config.getPushRowFilter() == null ||
+        carbonTableReader.config.getPushRowFilter().equalsIgnoreCase("false");
+    return new CarbondataPageSource(
+        carbonTable, queryId, carbonSplit, columns, configuration, isDirectVectorFill);
   }
 
   /**
