+++ integration/presto/src/main/java/org/apache/carbondata/presto/impl/CarbonLocalInputSplit.java
@@ -22,6 +22,7 @@ import java.util.List;
 
 import org.apache.carbondata.core.indexstore.BlockletDetailInfo;
 import org.apache.carbondata.core.metadata.ColumnarFormatVersion;
+import org.apache.carbondata.core.statusmanager.FileFormat;
 import org.apache.carbondata.hadoop.CarbonInputSplit;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
@@ -44,9 +45,9 @@ public class CarbonLocalInputSplit {
   private short version;
   private String[] deleteDeltaFiles;
   private String blockletId;
-
-
   private String detailInfo;
+  private int fileFormatOrdinal;
+  private FileFormat fileFormat;
 
   /**
    * Number of BlockLets in a block
@@ -93,6 +94,14 @@ public class CarbonLocalInputSplit {
     return blockletId;
   }
 
+  @JsonProperty public int getFileFormatOrdinal() {
+    return fileFormatOrdinal;
+  }
+
+  public FileFormat getFileFormat() {
+    return fileFormat;
+  }
+
   public void setDetailInfo(BlockletDetailInfo blockletDetailInfo) {
     Gson gson = new Gson();
     detailInfo = gson.toJson(blockletDetailInfo);
@@ -107,7 +116,8 @@ public class CarbonLocalInputSplit {
       @JsonProperty("version") short version,
       @JsonProperty("deleteDeltaFiles") String[] deleteDeltaFiles,
       @JsonProperty("blockletId") String blockletId,
-      @JsonProperty("detailInfo") String detailInfo
+      @JsonProperty("detailInfo") String detailInfo,
+      @JsonProperty("fileFormatOrdinal") int fileFormatOrdinal
   ) {
     this.path = path;
     this.start = start;
@@ -120,7 +130,8 @@ public class CarbonLocalInputSplit {
     this.deleteDeltaFiles = deleteDeltaFiles;
     this.blockletId = blockletId;
     this.detailInfo = detailInfo;
-
+    this.fileFormatOrdinal = fileFormatOrdinal;
+    this.fileFormat = FileFormat.getByOrdinal(fileFormatOrdinal);
   }
 
   public static CarbonInputSplit convertSplit(CarbonLocalInputSplit carbonLocalInputSplit) {
@@ -132,18 +143,21 @@ public class CarbonLocalInputSplit {
         carbonLocalInputSplit.getNumberOfBlocklets(),
         ColumnarFormatVersion.valueOf(carbonLocalInputSplit.getVersion()),
         carbonLocalInputSplit.getDeleteDeltaFiles());
-    Gson gson = new Gson();
-    BlockletDetailInfo blockletDetailInfo =
-        gson.fromJson(carbonLocalInputSplit.detailInfo, BlockletDetailInfo.class);
-    if (null == blockletDetailInfo) {
-      throw new RuntimeException("Could not read blocklet details");
+    inputSplit.setFormat(carbonLocalInputSplit.getFileFormat());
+    if (FileFormat.COLUMNAR_V3.ordinal() == inputSplit.getFileFormat().ordinal()) {
+      Gson gson = new Gson();
+      BlockletDetailInfo blockletDetailInfo =
+          gson.fromJson(carbonLocalInputSplit.detailInfo, BlockletDetailInfo.class);
+      if (null == blockletDetailInfo) {
+        throw new RuntimeException("Could not read blocklet details");
+      }
+      try {
+        blockletDetailInfo.readColumnSchema(blockletDetailInfo.getColumnSchemaBinary());
+      } catch (IOException e) {
+        throw new RuntimeException(e);
+      }
+      inputSplit.setDetailInfo(blockletDetailInfo);
     }
-    try {
-      blockletDetailInfo.readColumnSchema(blockletDetailInfo.getColumnSchemaBinary());
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-    inputSplit.setDetailInfo(blockletDetailInfo);
     return inputSplit;
   }
